# src/web_server/docs/swagger.yaml
definitions:
  models.BaseErr:
    properties:
      Message:
        example: 错误描述信息
        type: string
      code:
        example: 错误代码
        type: string
    type: object
  models.BaseResp:
    properties:
      RequestID:
        example: 请求ID
        type: string
      data:
        additionalProperties:
          type: string
        example:
          count: 数据条数
          info: 数据数组对象
        type: object
      error:
        $ref: '#/definitions/models.BaseErr'
    type: object
  models.ExtendInstFilter:
    properties:
      field:
        example: 模型属性名
        type: string
      value:
        example: 对应数据
        type: string
    type: object
  models.Issue:
    properties:
      create_time:
        type: string
      data:
        additionalProperties:
          type: string
        description: map[string]interface{}
        example:
          key: value
          key2: value2
        type: object
      data1:
        description: '[]string'
        example:
        - value1
        - value2
        items:
          type: string
        type: array
      field:
        example: 这是描述
        type: string
      id:
        type: integer
      solved_user:
        example: 这是解决人
        type: string
      value:
        example: 这是解决方法
        type: string
    type: object
  models.Result:
    properties:
      data: {}
      message:
        example: 请求信息
        type: string
      stat:
        example: 0
        type: integer
    type: object
info:
  contact: {}
paths:
  /issue/add:
    post:
      consumes:
      - application/json
      description: 创建issue desc
      parameters:
      - description: issue params
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Issue'
      - description: 人名
        in: query
        name: who
        required: true
        type: string
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: admin
        in: header
        name: BK_NAME
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功后返回
          headers:
            Token:
              description: token
              type: string
          schema:
            $ref: '#/definitions/models.Result'
      summary: 创建issue summary
      tags:
      - 创建issue summary tag
  /issue/list:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Issue'
                  type: array
              type: object
      summary: issue列表
      tags:
      - issue
  /lman-cmdb/v3/find/instassociation/object/:objectID:
    get:
      consumes:
      - application/json
      description: urlPath objectID 模型名称
      parameters:
      - description: request params
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.ExtendInstFilter'
      produces:
      - application/json
      responses:
        "200":
          description: 成功后返回
          headers:
            Token:
              description: token
              type: string
          schema:
            $ref: '#/definitions/models.BaseResp'
      summary: 按条件查询模型实例
      tags:
      - 按条件查询模型实例接口
swagger: "2.0"